@@ -1,6 +1,7 @@
+import { BrowserAnimationsModule } from '@angular/platform-browser/animations';
 import { NgModule, CUSTOM_ELEMENTS_SCHEMA, NO_ERRORS_SCHEMA } from '@angular/core';
 import { CommonModule } from '@angular/common';
-import { MatButtonModule, MatToolbarModule, MatProgressSpinnerModule, MatCardModule } from '@angular/material';
+import { MatButtonModule, MatToolbarModule, MatProgressSpinnerModule, MatCardModule, MatGridListModule } from '@angular/material';
 import { MatInputModule, MatListModule, MatDividerModule } from '@angular/material';
 
 @NgModule({
@@ -15,7 +16,9 @@ import { MatInputModule, MatListModule, MatDividerModule } from '@angular/materi
     MatProgressSpinnerModule,
     MatInputModule,
     MatListModule,
-    MatDividerModule
+    MatDividerModule,
+    MatGridListModule,
+    BrowserAnimationsModule
   ],
   declarations: [],
   schemas: [NO_ERRORS_SCHEMA]
