@@ -1,4 +1,7 @@
+import { MaterialModule } from './../material.module';
 import { async, ComponentFixture, TestBed } from '@angular/core/testing';
+import { FormsModule } from '@angular/forms';
+import { UppyfileModule } from './../uppyfile.module';
 
 import { UtilitiesComponent } from './utilities.component';
 
@@ -8,7 +11,12 @@ describe('UtilitiesComponent', () => {
 
   beforeEach(async(() => {
     TestBed.configureTestingModule({
-      declarations: [ UtilitiesComponent ]
+      declarations: [ UtilitiesComponent ],
+      imports: [
+        MaterialModule,
+        UppyfileModule,
+        FormsModule
+      ]
     })
     .compileComponents();
   }));
